requirements.txt
Django==2.1.4
django-cors-headers==2.4.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
mixer==6.1.3
pytest==4.0.1
pytest-cov==2.6.0
pytest-django==3.4.4
graphene_django==2.2.0